GraphBug-AI-Service/requirements.txt
neo4j>=5.15.0,<6.0.0
qdrant-client>=1.7.0,<2.0.0
tree-sitter==0.20.4
tree-sitter-languages==1.10.2
openai>=1.0.0,<2.0.0
numpy>=1.24.0,<2.0.0
sentence-transformers>=2.2.0,<3.0.0
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
gitpython>=3.1.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
langgraph>=0.2.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-google-genai>=2.0.0,<3.0.0
google-genai>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
PyGithub>=2.1.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
cryptography>=41.0.0,<43.0.0
httpx>=0.27.0,<1.0.0
slowapi>=0.1.9,<1.0.0